Move dark-mode hook to separate module
