+++ src_js/components/main_content/index.tsx
@@ -2,7 +2,7 @@ import { h } from 'preact';
 import { useCallback, useEffect, useRef } from 'preact/hooks';
 import clsx from 'clsx';
 import Config from '../../Config';
-import { usePrintInProgress } from '../../utils/hooks';
+import { usePrintInProgress } from '../../utils/hooks/print';
 import useTaskListCheckboxes from './useTaskListCheckboxes';
 
 type PropsType = {
