+++ src_js/utils/hooks/print.ts
@@ -94,23 +94,3 @@ export function useAfterPrint(handler: () => void): () => void {
     window.removeEventListener('afterprint', handler);
   };
 }
-
-export function usePrefersDarkMode(): boolean {
-  const [prefersDarkMode, setPrefersDarkMode] = useState(
-    window.matchMedia('(prefers-color-scheme: dark)').matches,
-  );
-  useEffect(() => {
-    const listener = (mql: MediaQueryListEvent) =>
-      setPrefersDarkMode(mql.matches);
-
-    window
-      .matchMedia('(prefers-color-scheme: dark)')
-      .addEventListener('change', listener);
-    return () => {
-      window
-        .matchMedia('(prefers-color-scheme: dark)')
-        .removeEventListener('change', listener);
-    };
-  }, []);
-  return prefersDarkMode;
-}
